# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  devops-cd_internal_dto.CreateBatchApp:
    properties:
      app_id:
        description: 应用ID
        type: integer
      release_notes:
        description: 应用级发布说明（可选）
        type: string
    required:
    - app_id
    type: object
  devops-cd_internal_dto.CreateBatchRequest:
    properties:
      apps:
        description: 应用列表（允许为空，封板时校验）
        items:
          $ref: '#/definitions/devops-cd_internal_dto.CreateBatchApp'
        type: array
      batch_number:
        description: 批次编号/标题，用户填写
        type: string
      project_id:
        description: 关联的项目ID
        type: integer
      release_notes:
        description: 批次级发布说明（可选）
        type: string
    required:
    - batch_number
    - project_id
    type: object
...
  devops-cd_internal_dto.UpdateBatchRequest:
    properties:
      add_apps:
        description: 新增应用
        items:
          $ref: '#/definitions/devops-cd_internal_dto.CreateBatchApp'
        type: array
      batch_id:
        type: integer
      batch_number:
        description: Operator     string           `json:"operator" binding:"required"`
        type: string
      release_notes:
        type: string
      remove_app_ids:
        items:
          type: integer
        type: array
    required:
    - batch_id
    type: object
...
  internal_api_handler.ApproveRequest:
    properties:
      batch_id:
        description: 批次ID
        type: integer
      operator:
        description: 审核人
        type: string
      reason:
        description: 审核意见
        type: string
    required:
    - batch_id
    - operator
    type: object
  internal_api_handler.ProcessActionRequest:
    properties:
      action:
        description: '操作类型: seal/start_pre_deploy/finish_pre_deploy/start_prod_deploy/finish_prod_deploy/complete/cancel'
        type: string
      batch_id:
        description: 批次ID
        type: integer
      operator:
        description: 操作人
        type: string
      reason:
        description: 原因（可选）
        type: string
    required:
    - action
    - batch_id
    - operator
    type: object
  internal_api_handler.RejectRequest:
    properties:
      batch_id:
        description: 批次ID
        type: integer
      operator:
        description: 审核人
        type: string
      reason:
        description: 拒绝原因
        type: string
    required:
    - batch_id
    - operator
    - reason
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    DevOps 持续交付平台 API 文档
    提供代码库管理、应用管理、批次管理、部署管理等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: DevOps CD API
  version: "1.0"
paths:
...
  /api/v1/batch:
    get:
      consumes:
      - application/json
      description: 获取批次详细信息和应用列表（支持应用列表分页）。每个应用默认会包含自上次部署以来的最近15条成功构建记录，可通过 with_recent_builds=false
        关闭。
      parameters:
      - description: 批次ID
        format: int64
        in: query
        name: id
        required: true
        type: integer
      - default: 1
        description: 应用列表页码
        in: query
        name: app_page
        type: integer
      - default: 20
        description: 应用列表每页数量
        in: query
        name: app_page_size
        type: integer
      - default: true
        description: 是否包含最近构建记录
        in: query
        name: with_recent_builds
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取批次详情
      tags:
      - 批次管理
    post:
      consumes:
      - application/json
      description: 创建新的发布批次，自动扫描指定应用的最新成功构建
      parameters:
      - description: 批次创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.CreateBatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "409":
          description: 应用冲突
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建批次
      tags:
      - 批次管理
    put:
      consumes:
      - application/json
      description: 更新批次基本信息、添加或删除应用
      parameters:
      - description: 批次更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.UpdateBatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 批次已封板
          schema:
            additionalProperties: true
            type: object
        "409":
          description: 应用冲突
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新批次
      tags:
      - 批次管理
  /api/v1/batch/action:
    post:
      consumes:
      - application/json
      description: '处理批次的各种状态操作，如封板、部署、验收等。支持的action: seal(封板)/start_pre_deploy(开始预发布)/finish_pre_deploy(完成预发布)/start_prod_deploy(开始生产部署)/finish_prod_deploy(完成生产部署)/complete(最终验收)/cancel(取消)'
      parameters:
      - description: 操作请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.ProcessActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 操作失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 批次状态操作
      tags:
      - 批次管理
  /api/v1/batch/approve:
    post:
      consumes:
      - application/json
      description: 审批通过批次，更新审批状态为approved
      parameters:
      - description: 审批请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.ApproveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 审批失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 批次审批通过
      tags:
      - 批次管理
  /api/v1/batch/reject:
    post:
      consumes:
      - application/json
      description: 审批拒绝批次，更新审批状态为rejected
      parameters:
      - description: 拒绝请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.RejectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 拒绝失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 批次审批拒绝
      tags:
      - 批次管理
...
  /api/v1/batch/status:
    get:
      consumes:
      - application/json
      description: 获取批次状态信息（轻量级接口，专门用于状态轮询）。只返回批次和应用的状态信息，不包含构建历史、依赖等详细数据。
      parameters:
      - description: 批次ID
        format: int64
        in: query
        name: id
        required: true
        type: integer
      - default: 1
        description: 应用列表页码
        in: query
        name: app_page
        type: integer
      - default: 20
        description: 应用列表每页数量
        in: query
        name: app_page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取批次状态
      tags:
      - 批次管理
  /api/v1/batches:
    get:
      consumes:
      - application/json
      description: 分页查询批次列表，支持状态、发起人、审批状态、时间范围、关键字过滤。status支持多值，例如：?status=1&status=2&status=3
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - collectionFormat: csv
        description: 状态过滤（支持多个）
        in: query
        items:
          format: int32
          type: integer
        name: status
        type: array
      - description: 发起人过滤
        in: query
        name: initiator
        type: string
      - description: 审批状态过滤：pending/approved/rejected
        in: query
        name: approval_status
        type: string
      - description: 创建时间起始（RFC3339格式）
        in: query
        name: created_at_start
        type: string
      - description: 创建时间结束（RFC3339格式）
        in: query
        name: created_at_end
        type: string
      - description: 关键字搜索（批次编号、发起人、发布说明）
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 查询批次列表
      tags:
      - 批次管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"